chore: dev cleartext
