+++ .github/workflows/android-weekly.yml
@@ -87,6 +87,7 @@ jobs:
       - name: python fixHTTP
         run: |
           python fixHTTP.py
+          python ./scripts/dev_cleartext.py
 
       - name: WGet Build Sign jks
         uses: wei/wget@v1
