+++ scripts/dev_cleartext.py
@@ -1,0 +1,18 @@
+from pathlib import Path
+import os
+
+
+def fix_content(path: str, transform) -> None:
+    content = []
+    with open(Path(path), encoding='utf8') as f:
+        for i in f:
+            content.append(transform(i))
+
+    with open(Path(path), 'w', encoding='utf8') as f:
+        for line in content:
+            f.write(line)
+
+
+if __name__ == '__main__':
+    fix_content('android/app/src/main/AndroidManifest.xml', lambda line: line.replace(
+        'android:usesCleartextTraffic="false"', 'android:usesCleartextTraffic="true"'))
